Initial commit - Old sources recovery
This is a small handwritten driver for the MAXIM MAX7219 chips (used along
with 7-segments displays) on the Raspberry-Pi.
It was functional at the time I wrote it but I couldn't get it to work on
a more recent version of Raspbian, sadly.
Some updates might come later when I get a newer RPi at home to tinker
with.